add early stop generation.

+++ generate.py
@@ -103,7 +103,7 @@ class LmGeneration:
         continue_exsample = [i for i in range(batch)]
         with torch.no_grad():
             for cur_pos in range(start_pos, total_len):
-                logits = self.model.forward(tokens[continue_exsample, prev_pos:cur_pos], prev_pos, continue_exsample)
+                logits = self.model.forward(tokens[continue_exsample, prev_pos:cur_pos], prev_pos, continue_exsample).float()
                 if args.temperature > 0:
                     #next_token_scores = top_k_top_p_filtering(logits, top_k=args.top_k, top_p=args.top_p)
                     next_token_scores = apply_top_p(logits, args.top_p)
